+++ sources/x/create.tardigrade.ts
@@ -9,10 +9,10 @@ const sessionKey: symbol = genSessionKey();
 console.log('Tardigrade v1.2.0');
 /// processing-version >>
 
-export const createTardigrade = (initialData?: Dictionary, initialOptions?: TardigradeInitialOptions): Tardigrade => {
+export const createTardigrade = <S extends Dictionary = Dictionary>(initialData?: S, initialOptions?: TardigradeInitialOptions): Tardigrade<S> => {
     initialOptions = initialOptions || {};
 
-    const instance: Tardigrade = new Tardigrade(sessionKey, initialOptions);
+    const instance: Tardigrade<S> = new Tardigrade<S>(sessionKey, initialOptions);
 
     if (!initialData) {
         return instance;